backend/requirements.txt
aiohttp
pyautogui
opencv-python
numpy
websockets
pytesseract
python-dotenv
google-generativeai
pillow
gtts
pymongo[srv]
PyJWT
bcrypt
motor
langchain_community
langchain_google_genai
langchain
pypdf
pymilvus
aiohttp_cors